Moved calendar class to own file
